refactor: move Newton-Rhapson methods to a new nr module
