+++ src/math/mod.rs
@@ -1,6 +1,4 @@
-use std::hint;
-
-use nalgebra::{Complex, Dim, OVector, SMatrix, SVector};
+use nalgebra::{Complex, Dim, OVector};
 use num_traits::Float;
 use numeric_literals::replace_float_literals;
 use simba::simd::{SimdBool, SimdComplexField};
@@ -9,119 +7,9 @@ use crate::Scalar;
 
 pub mod interpolation;
 pub mod lut;
+pub mod nr;
 #[cfg(feature = "math-polynom")]
 pub mod polynom;
-
-/// Trait desciring a multivariate root equation. Root equations are solved with numerical methods such as
-/// Newton-Rhapson, when linear algebra cannot be used (e.g. in the case of nonlinear systems).
-pub trait RootEq<T, const N: usize> {
-    /// Evaluate the equation at the given input vector.
-    fn eval(&self, input: &SVector<T, N>) -> SVector<T, N>;
-
-    /// Evaluate the **inverse** jacobian of the equation. When the jacobian cannot be computed, [`None`] should be returned instead.
-    fn j_inv(&self, input: &SVector<T, N>) -> Option<SMatrix<T, N, N>>;
-}
-
-/// Perform a single step of the Newton-Rhapson algorithm. This takes the inverse jabobian and computes the differential to the next step.
-#[cfg_attr(test, inline(never))]
-#[cfg_attr(not(test), inline)]
-pub fn nr_step<T: Scalar, const N: usize>(
-    eq: &impl RootEq<T, N>,
-    input: &SVector<T, N>,
-) -> Option<SVector<T, N>>
-where
-    T::Element: Float,
-{
-    let ret = eq.j_inv(input).map(|jinv| jinv * eq.eval(input))?;
-    let all_finite = ret
-        .iter()
-        .copied()
-        .flat_map(|v| v.into_iter())
-        .all(|v| v.is_finite());
-
-    all_finite.then_some(ret)
-}
-
-/// Solve the given root equation using Newton-Rhapson for a specified number of setps.
-#[inline]
-#[profiling::function]
-pub fn newton_rhapson_steps<T: Scalar, const N: usize>(
-    eq: &impl RootEq<T, N>,
-    value: &mut SVector<T, N>,
-    iter: usize,
-) -> T
-where
-    T::Element: Float,
-{
-    let Some(mut step) = nr_step(eq, value) else {
-        return T::zero();
-    };
-    for _ in 1..iter {
-        *value -= step;
-        if let Some(newstep) = nr_step(eq, value) {
-            step = newstep;
-        } else {
-            break;
-        }
-    }
-    rms(&step)
-}
-
-/// Solve the given root equation using Newton-Rhapson until the RMS of the differential is lesser than the given tolerance.
-#[cfg_attr(test, inline(never))]
-#[cfg_attr(not(test), inline)]
-#[profiling::function]
-pub fn newton_rhapson_tolerance<T: Scalar, const N: usize>(
-    eq: &impl RootEq<T, N>,
-    value: &mut SVector<T, N>,
-    tol: T,
-) -> usize
-where
-    T::Element: Float,
-{
-    let mut i = 0;
-
-    while let Some(step) = nr_step(eq, value) {
-        debug_assert!(i < 500, "Newton-Rhapson got stuck");
-        i += 1;
-        let tgt = hint::black_box(rms(&step)).simd_lt(tol);
-        if tgt.all() {
-            break;
-        }
-        let changed = *value - step;
-        *value = value.zip_map(&changed, |v, c| v.select(tgt, c));
-    }
-
-    i
-}
-
-/// Solve the given root equation using Newton-Rhapson, until either the RMS of the differential is less than the
-/// given tolerances, or the specified max number of steps has been taken.
-#[cfg_attr(test, inline(never))]
-#[cfg_attr(not(test), inline)]
-#[profiling::function]
-pub fn newton_rhapson_tol_max_iter<T: Scalar, const N: usize>(
-    eq: &impl RootEq<T, N>,
-    value: &mut SVector<T, N>,
-    tol: T,
-    max_iter: usize,
-) -> usize
-where
-    T::Element: Float,
-{
-    for i in 0..max_iter {
-        let Some(step) = nr_step(eq, value) else {
-            return i;
-        };
-        let changed = *value - step;
-        let tgt = rms(&step).simd_lt(tol);
-        *value = value.zip_map(&changed, |v, c| v.select(tgt, c));
-        if tgt.all() {
-            return i;
-        }
-    }
-    return max_iter;
-}
 
 #[replace_float_literals(Complex::from(T::from_f64(literal)))]
 pub fn freq_to_z<T: Scalar>(samplerate: T, f: T) -> Complex<T>
@@ -171,58 +59,4 @@ pub fn smooth_max<T: Scalar>(t: T, a: T, b: T) -> T {
 #[inline]
 pub fn smooth_clamp<T: Scalar>(t: T, x: T, min: T, max: T) -> T {
     smooth_max(t, min, smooth_min(t, x, max))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    struct SqrtNumerical {
-        squared: f64,
-    }
-
-    impl RootEq<f64, 1> for SqrtNumerical {
-        fn eval(&self, input: &SVector<f64, 1>) -> SVector<f64, 1> {
-            SVector::<_, 1>::new(self.squared - input[0].powi(2))
-        }
-
-        fn j_inv(&self, input: &SVector<f64, 1>) -> Option<SMatrix<f64, 1, 1>> {
-            let diff = -2.0 * input[0];
-            Some(SVector::<_, 1>::new((diff + 1e-6).recip()))
-        }
-    }
-
-    #[test]
-    fn test_solve() {
-        let equ = SqrtNumerical { squared: 4.0 };
-        let mut actual = SVector::<_, 1>::new(0.0);
-        let iters = newton_rhapson_tolerance(&equ, &mut actual, 1e-4);
-        let expected = 2.0;
-
-        println!("Iterations: {iters} | {actual}");
-        assert!((expected - actual[0].abs()).abs() <= 1e-4);
-    }
-
-    #[test]
-    fn test_detect_nan() {
-        struct Equ;
-
-        impl RootEq<f64, 1> for Equ {
-            fn eval(&self, _: &SVector<f64, 1>) -> SVector<f64, 1> {
-                SVector::zeros()
-            }
-
-            fn j_inv(&self, _: &SVector<f64, 1>) -> Option<SMatrix<f64, 1, 1>> {
-                Some(SVector::<_, 1>::new(f64::NAN))
-            }
-        }
-
-        assert_eq!(
-            0,
-            newton_rhapson_tolerance(&Equ, &mut SVector::zeros(), 1e-4)
-        );
-    }
-
-    #[test]
-    fn test_freq_to_z() {}
 }
